+++ PasosIniciales/Models/Producto3.cs
@@ -4,7 +4,7 @@
     {
         public string Id { get; set; }
         public string Nombre { get; set; }
-        public double Precio { get; set; }
+        public Decimal Precio { get; set; }
         public string? Foto { get; set; }
     }
 }
